Remove dead ratchet review-poll subsystem and relocate reconciliation
Zero-risk ratchet cleanups from an architecture audit:

- Delete the review-comment poll subsystem: its re-fetch always ran
  milliseconds after the main check registered a fetch in the 90s-cooldown
  PR fetch registry, so it unconditionally short-circuited to 'waiting'
  and the comments-found path was unreachable. The main 2-minute poll
  already detects new review activity via the dispatch snapshot key.
- Remove clearWorkspaceState (only purpose was clearing poll trackers)
  from the ratchet service and the workspace archive orchestrator.
- Move reconciliation.service.ts to orchestration/: it provisions
  worktrees and cleans orphaned terminal sessions, which is cross-service
  coordination unrelated to ratcheting. Drop the now-unused
  RatchetWorkspaceBridge and the ratchet->terminal registry dependency.
- Stop polling closed PRs: findWithPRsForRatchet now excludes
  prState CLOSED (previously fetched from GitHub every 2 minutes forever;
  the scheduler PR sync still refreshes prState, so reopened PRs re-enter).
- Document that ratchetLastCiRunId actually stores the dispatch snapshot
  key, not a CI run id.
