+++ src/backend/orchestration/workspace-archive.orchestrator.ts
@@ -27,9 +27,6 @@ export type ArchiveWorkspaceDependencies = {
     stopRunScript(workspaceId: string): Promise<{ success: boolean; error?: string }>;
     evictWorkspaceBuffers(workspaceId: string): void;
   };
-  ratchetService: {
-    clearWorkspaceState(workspaceId: string): void;
-  };
   sessionService: {
     stopWorkspaceSessions(workspaceId: string): Promise<void>;
   };
@@ -148,7 +145,6 @@ async function completeArchive(
 
   const archivedWorkspace = await workspaceStateMachine.markArchived(workspace.id);
   services.runScriptService.evictWorkspaceBuffers(workspace.id);
-  services.ratchetService.clearWorkspaceState(workspace.id);
 
   // Notify parent workspace that this child was archived (fire-and-forget)
   fireLifecycleNotification(
